requirements.txt
requests
boto3
azure-storage-blob
pre-commit
black
psutil
